Refactored

+++ modules/siddhi-extensions/reorder/src/main/java/org/wso2/siddhi/extension/reorder/alphakslack/Runtime.java
@@ -58,7 +58,7 @@ public class Runtime{
 
                 if ((cdistance < distance) && (cdistance != 0)) {
                     distance = cdistance;
-                    lowerIndex = (long) count;
+                    lowerIndex = count;
                 }
                 count += 1;
             }
